Add files via upload
Create a java class LastnameFirstnameA2.java to store an integer array and a number. The output would be all the pairs of integers from the array which add up to that number.

Sample input:
Array = [2, 5, 6, 8, -3, -2, 10, 7, 11, 4]
Num = 8

Sample output:
Array = [2, 5, 6, 8, -3, -2, 10, 7, 11, 4]
No_of_elements =  10
Sum = 8
(2, 6)
(-2, 10)
(-3, 11)

The output shows the array, the number as the ‘Sum’ and all the pairs of integers adding to that number.