Add files via upload
“Lotter Game”

Write a program that generates a random integer in the range 25 to 75 inclusive. Accept an integer in the same range from the user playing lottery and check if it equals the random number generated at the first step. The player/user gets 3 chances to guess the random number and the program must quit at the trial when the user wins or when user exhausts the 3 trials. At the end of 3 trials or at the trial when the user wins, display the random number that was generated, number of trials and the message if user won the lottery or lost it.
Write a class for the game named as LastNameFirstnameA5.java and define a method called playGame() which will be called from the main method in the test class we create.
Sample output:
This is your trial 1. Guess a number to win the lottery!
34
Alas! Please try again. This is your trial 2!
28
Bad luck! You got one more trial!
39
Sorry! You lost!
Random number generated: 55
Number of trials: 3
Would you like to play again? (Y/N) Y
This is your trial 1. Guess a number to win the lottery!
26
Alas! Please try again. This is your trial 2!
37
Hurray! You won the lottery!
Random number generated: 37
Number of trials: 2
Would you like to play again? (Y/N) N